+++ pyp5js/config/fs.py
@@ -18,12 +18,8 @@ class LibFiles():
         return self.install.joinpath('templates')
 
     @property
-    def assets_dir(self):
-        return self.install.joinpath('assets')
-
-    @property
     def static_dir(self):
-        return self.assets_dir.joinpath('static')
+        return self.install.joinpath('http', 'static')
 
     @property
     def pytop5js(self):
@@ -31,11 +27,11 @@ class LibFiles():
 
     @property
     def p5js(self):
-        return self.static_dir.joinpath('p5', 'p5.min.js')
+        return self.static_dir.joinpath('js', 'p5', 'p5.min.js')
 
     @property
     def p5_yml(self):
-        return self.assets_dir.joinpath('p5_reference.yml')
+        return self.static_dir.joinpath('p5_reference.yml')
 
     ##### TRANSCRYPT SPECIFICS