+++ pyp5js/tests/test_templates_renderers.py
@@ -5,13 +5,14 @@ from pyp5js.sketch import Sketch
 from pyp5js.config.fs import PYP5JS_FILES
 
 from .fixtures import sketch
+from ..config.sketch import P5_JS_CDN
 
 
 def test_get_sketch_index_content(sketch):
     expected_template = renderers.templates.get_template('transcrypt/index.html')
     expected_content = expected_template.render({
         'sketch_name': sketch.sketch_name,
-        "p5_js_url": sketch.STATIC_NAME + "/p5.js",
+        "p5_js_url": P5_JS_CDN,
         "sketch_js_url": sketch.TARGET_NAME + "/target_sketch.js",
     })