+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 Development
 -----------
+- Serve pyodide JS files locally [PR #186](https://github.com/berinhard/pyp5js/pull/186)
+- Create sketch using p5.js from CDN [PR #191](https://github.com/berinhard/pyp5js/pull/191)
 - Add `keyIsDown` event to Transcrypt mode - [PR #187](https://github.com/berinhard/pyp5js/pull/187)
 - Fix bug with multiple events calls - PR #187 too